Rename `Utils` to `EquoChromium` and shrink the API.
