+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/EquoIdeTask.java
@@ -14,8 +14,8 @@
 package dev.equo.ide.gradle;
 
 import dev.equo.ide.BuildPluginIdeMain;
+import dev.equo.ide.EquoChromium;
 import dev.equo.ide.IdeHook;
-import dev.equo.ide.chromium.Utils;
 import dev.equo.solstice.p2.P2QueryResult;
 import java.io.File;
 import java.io.IOException;
@@ -140,7 +140,7 @@ public abstract class EquoIdeTask extends DefaultTask {
 		}
 
 		if (P2ModelDsl.isChromiumEnabled()) {
-			Utils.removeSwtSigner(classpath);
+			EquoChromium.removeSwtSigner(classpath);
 		}
 
 		caller.ideHooks = ideHooks;
